stock is separated from resource entity, recording a sale reflects updated stock and weighted avg on resource list

+++ wowStockCalculator/src/main/java/com/david/wowStockCalculator/domain/entities/Sale.java
@@ -30,7 +30,7 @@ public class Sale {
 
     private Integer amount;
 
-    private Integer cost;
+    private Long cost;
 
     @Column(name = "isSold", columnDefinition = "boolean default false")
     private Boolean isSold = false;
